Refactor: train_trad and train_nn created
